+++ apps/main-landing/src/components/products-section/components/creators-section-actions.tsx
@@ -1,0 +1,16 @@
+import { Button } from '@idriss-xyz/ui/button';
+export const CreatorsSectionActions = () => {
+  return (
+    <Button
+      intent="secondary"
+      size="large"
+      className="w-full text-button2 lg:w-fit lg:text-button1"
+      asLink
+      href="https://chromewebstore.google.com/detail/idriss/fghhpjoffbgecjikiipbkpdakfmkbmig"
+      target="_blank"
+      rel="noopener noreferrer"
+    >
+      CREATE DONATION LINK
+    </Button>
+  );
+};
