+++ apps/main-landing/src/components/products-section/components/creators-section.tsx
@@ -1,8 +1,9 @@
 'use client';
-import { Button } from '@idriss-xyz/ui/button';
+
 import { ProductSection } from './product-section';
 import { CREATORS_INFO } from '../constants';
 import { TabOption } from '../types';
+import { CreatorsSectionActions } from './creators-section-actions';
 
 type Properties = {
   onOptionChange: (option: TabOption) => void;
@@ -19,20 +20,8 @@ export const CreatorsSection = ({
     <ProductSection
       fadeOut={fadeOut}
       className={className}
-      actions={
-        <Button
-          intent="secondary"
-          size="large"
-          className="text-button2 lg:text-button1"
-          asLink
-          href="https://chromewebstore.google.com/detail/idriss/fghhpjoffbgecjikiipbkpdakfmkbmig"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          CREATE DONATION LINK
-        </Button>
-      }
-      defaultOptionKey="creators-product"
+      actions={<CreatorsSectionActions />}
+      activeOptionKey="creators-product"
       description={CREATORS_INFO.describtion}
       title={CREATORS_INFO.title}
       features={CREATORS_INFO.features}
