+++ apps/main-landing/src/components/products-section/components/prediction-markets-section.tsx
@@ -1,8 +1,8 @@
 'use client';
-import { Button } from '@idriss-xyz/ui/button';
 import { ProductSection } from './product-section';
 import { PREDICTION_MARKETS_INFO } from '../constants';
 import { TabOption } from '../types';
+import { PredictionMarketsSectionActions } from './prediction-markets-section-actions';
 
 type Properties = {
   onOptionChange: (option: TabOption) => void;
@@ -19,26 +19,8 @@ export const PredictionMarketsSection = ({
     <ProductSection
       fadeOut={fadeOut}
       className={className}
-      actions={
-        <>
-          <input
-            className="flex h-12 w-[290px] flex-[1_0_0] items-center rounded-[12px] border px-3 py-2 shadow-[0_0_0_4px_rgba(242,242,242,0.14)]"
-            placeholder="Your email"
-          />
-          <Button
-            intent="secondary"
-            size="medium"
-            className="text-button2 lg:text-button1"
-            asLink
-            href="https://chromewebstore.google.com/detail/idriss/fghhpjoffbgecjikiipbkpdakfmkbmig"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            GET EARLY ACCESS
-          </Button>
-        </>
-      }
-      defaultOptionKey="prediction-markets-product"
+      actions={<PredictionMarketsSectionActions />}
+      activeOptionKey="prediction-markets-product"
       description={PREDICTION_MARKETS_INFO.describtion}
       title={PREDICTION_MARKETS_INFO.title}
       features={PREDICTION_MARKETS_INFO.features}
